+++ packages/toolkit/src/utils/getAragonArtifact.ts
@@ -11,11 +11,6 @@ export default function getAragonArtifact(
   functions: AragonContractFunction[],
   abi: AbiItem[]
 ): AragonArtifact {
-  const mainnetEnvName = 'mainnet'
-  const mainnetEnvironment = arapp.environments[mainnetEnvName]
-  const mainnetAppName = (mainnetEnvironment || {}).appName || ''
-  const mainnetAppId = mainnetAppName ? namehash.hash(mainnetAppName) : ''
-
   return {
     ...arapp,
 
@@ -48,10 +43,5 @@ export default function getAragonArtifact(
     abi,
     // Additional metadata
     flattenedCode: `./code.sol`,
-
-    // Curent environment, TODO: is this necessary?
-    env: mainnetEnvironment,
-    appName: mainnetAppName,
-    appId: mainnetAppId,
   }
 }
